+++ src/main/java/br/com/zalf/prolog/webservice/gente/prontuarioCondutor/model/ocorrencia/AcidentesTrabalho.java
@@ -3,7 +3,7 @@ package br.com.zalf.prolog.webservice.gente.prontuarioCondutor.model.ocorrencia;
 /**
  * Created by Zart on 03/07/2017.
  */
-public class AcidentesTrabalho {
+public final class AcidentesTrabalho extends Ocorrencia {
 
     private int fai;
     private int lti;