+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Change Log
 <a name="v1.0.6"></a>
 ## Version [v1.0.6](https://github.com/luizfp/PrologWebService/compare/v1.0.5...v1.0.6) (release-date) [unreleased]
 
+### Refactors
+* Adiciona informações de fadiga no prontuário do condutor (PL-2269)
+* Permite filtrar para incluir marcas não utilizadas na busca de marcas de pneus e bandas (PL-2390)
+* Cria métodos para import de planilha de veículos (PL-2318)
 
 <a name="v1.0.5"></a>
 ## Version [v1.0.5](https://github.com/luizfp/PrologWebService/compare/v1.0.4...v1.0.5) (2019-11-06)